Package restructuring + now ability to plot in MainFrame
