+++ pygedcom/gedcomparser/__init__.py
@@ -1,0 +1,1 @@
+from pygedcom.gedcomparser.parser import Parser
