+++ pygedcom/treeplot/__init__.py
@@ -1,0 +1,1 @@
+from pygedcom.treeplot.main import plot_tree
