+++ test.py
@@ -1,4 +1,4 @@
-from pygedcom.core import Parser
+from pygedcom.gedcomparser.parser import Parser
 
 file = 'data/555SAMPLE.GED'
 tree = Parser(file).build_tree()
